Create helpers and add center as a prop
